% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lssdoc-package.R
\docType{package}
\name{lssdoc-package}
\alias{lssdoc}
\alias{lssdoc-package}
\title{lssdoc: Render Multilingual Questionnaires from LimeSurvey '.lss' Files}
\description{
\if{html}{\figure{logo.svg}{options: style='float: right' alt='logo' width='120'}}

Render LimeSurvey '.lss' survey exports as multilingual questionnaire documents in Word ('.docx') or PDF, displaying up to four languages side by side with localized chrome in English, French, German, Spanish and Italian. Includes a rule-based automated audit that flags missing translations, forward filter references, duplicate codes, array-scale inconsistencies and orphan structural references. Designed for anyone working with a LimeSurvey survey: researchers, methodologists, ethics committees, translators and reviewers. Processing is fully local: the source file is the only input and no questionnaire content is uploaded to a third-party service.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://amaltawfik.github.io/lssdoc/}
  \item \url{https://github.com/amaltawfik/lssdoc}
  \item Report bugs at \url{https://github.com/amaltawfik/lssdoc/issues}
}

}
\author{
\strong{Maintainer}: Amal Tawfik \email{amal.tawfik@hesav.ch} (\href{https://orcid.org/0009-0006-2422-1555}{ORCID}) (\href{https://ror.org/04j47fz63}{ROR}) [copyright holder]

Authors:
\itemize{
  \item Amal Tawfik \email{amal.tawfik@hesav.ch} (\href{https://orcid.org/0009-0006-2422-1555}{ORCID}) (\href{https://ror.org/04j47fz63}{ROR}) [copyright holder]
}

}
\keyword{internal}
