% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lssdoc-package.R
\docType{package}
\name{lssdoc-package}
\alias{lssdoc}
\alias{lssdoc-package}
\title{lssdoc: Generate Word Review Documents from LimeSurvey '.lss' Files}
\description{
Build a professional Word ('.docx') review document from a LimeSurvey '.lss' export, displaying up to four languages side by side. The output supports human review of multilingual questionnaires by ethics committees, translators, and methodologists: it reproduces the questionnaire content faithfully and flags automatically detectable anomalies that do not require any AI. Processing is fully local, so no questionnaire is uploaded to any third-party service.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/amaltawfik/lssdoc}
  \item Report bugs at \url{https://github.com/amaltawfik/lssdoc/issues}
}

}
\author{
\strong{Maintainer}: Amal Tawfik \email{amal.tawfik@hesav.ch} (\href{https://orcid.org/0009-0006-2422-1555}{ORCID}) (\href{https://ror.org/04j47fz63}{ROR}) [copyright holder]

Authors:
\itemize{
  \item Amal Tawfik \email{amal.tawfik@hesav.ch} (\href{https://orcid.org/0009-0006-2422-1555}{ORCID}) (\href{https://ror.org/04j47fz63}{ROR}) [copyright holder]
}

}
\keyword{internal}
